Add endpoints. Add database migration "Remove SupplierId from Product"

+++ MartBerries-Server.API/Controllers/OrderController.cs
@@ -18,19 +18,19 @@ namespace MartBerries_Server.API.Controllers
         [HttpGet]
         public async Task<ActionResult<List<Order>>> Get()
         {
-            return await QueryAsync(new GetOrderListQuery());
+            return Single(await QueryAsync(new GetOrderListQuery()));
         }
 
         [HttpGet("{id:Guid}")]
         public async Task<ActionResult<Order>> Get(Guid id)
         {
-            return await QueryAsync(new GetOrderQuery(id));
+            return Single(await QueryAsync(new GetOrderQuery(id)));
         }
 
         [HttpGet("{statusId:int}")]
         public async Task<ActionResult<List<Order>>> GetByStatusId(int statusId)
         {
-            return await QueryAsync(new GetOrderListQuery(statusId));
+            return Single(await QueryAsync(new GetOrderListQuery(statusId)));
         }
 
         [HttpPost]
